+++ DataLayer/MyEntityDb/EfCompareDbs/DbQueryDbContext.cs
@@ -1,0 +1,23 @@
+﻿// Copyright (c) 2019 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
+// Licensed under MIT license. See License.txt in the project root for license information.
+
+using DataLayer.SpecialisedEntities;
+using Microsoft.EntityFrameworkCore;
+
+namespace DataLayer.MyEntityDb.EfCompareDbs
+{
+    public class DbQueryDbContext : DbContext
+    {
+        public DbSet<MyEntity> MyEntities { get; set; }
+        public DbQuery<MyEntityReadOnly> MyReadOnlyEntities { get; set; }
+
+        public DbQueryDbContext(
+            DbContextOptions<DbQueryDbContext> options)
+            : base(options) { }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.Entity<MyEntity>().ToTable(nameof(MyReadOnlyEntities));
+        }
+    }
+}